+++ src/parser/DocScriptFactory.ts
@@ -2,7 +2,7 @@ import DocExample from "../doc_models/DocExample";
 import DocParam from "../doc_models/DocParam";
 import DocReturns from "../doc_models/DocReturns";
 import DocScript from "../doc_models/DocScript";
-import StringUtils from "./StringUtils";
+import StringUtils from "../StringUtils";
 
 /**
  * Factory class to create DocScript instances.
@@ -90,7 +90,7 @@ export default class DocScriptFactory {
 	 */
 	public setDescription(descriptionText: string) {
 		if (descriptionText !== "") {
-			this._script.description = descriptionText;
+			this._script.description = StringUtils.stripInitialLineFeeds(descriptionText);
 			this._script.undocumented = false;
 		}
 	}
@@ -133,10 +133,8 @@ export default class DocScriptFactory {
 		param.name = StringUtils.escapeHtml(name);
 		param.type = type ? StringUtils.escapeHtml(type) : "";
 		param.optional = optional;
-		let str = StringUtils.stripInitialHypen(description);
-		str = StringUtils.markdown2Html(str);
-		str = StringUtils.compactHtmlSingleParagraph(str);
-		param.description = str;
+		description = StringUtils.stripInitialLineFeeds(description);
+		param.description = StringUtils.stripInitialHypen(description);
 		return param;
 	}
