+++ src/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../base-tsconfig.json",
+  "extends": "../tsconfig.json",
   "include": ["**/*.ts"],
   "compilerOptions": {
     "rootDir": ".",
