+++ tests/tsconfig.json
@@ -1,4 +1,4 @@
 {
-  "extends": "../base-tsconfig.json",
+  "extends": "../tsconfig.json",
   "include": ["**/*.ts"]
 }
